% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNMDS.R
\docType{package}
\name{fitNMDS-package}
\alias{fitNMDS-package}
\title{The fitNMDS Package for Evaluating Nonmetric
    Multidimensional Scaling (NMDS) Models}
\description{
Measures of agreement for bootstrapped and reciprocal
    NMDS.
}
\details{
Combining two different datasets into one nonmetric
    multidimensional scaling (NMDS) model can be risky if they each
    cover different attribute spaces (e.g., different species pools
    in ecology). Therefore, comparing two datasets requires
    estimating internal agreement (sampling variability) relative
    to external agreement (exchangeability).

    Bootstrapped NMDS with \code{boot_nmds} estimates internal
    agreement (sampling variability) of one candidate dataset.
    The premise of bootstrapped ordination is resampling with
    replacement from the reference dataset, performing NMDS
    ordination on each bootstrap replicate, then determining
    collective stability across all NMDS solutions.  Other uses of
    bootstrap NMDS may include estimating confidence regions for
    ordination site scores, or testing the stability of species
    positions along ordination axes.

    Reciprocal NMDS with \code{recip_nmds} estimates external
    agreement (exchangeability) of two candidate datasets.  The
    premise of reciprocal NMDS is to calibrate each of two models
    with each respective dataset, mutually exchanging datasets
    among the alternative ordination models, then determining
    measures of how well each dataset fit the alternative model.
    This is really just a special case of 2-fold cross-validation
    as applied to NMDS.  The primary use of reciprocal NMDS is to
    test the null hypothesis that two multivariate datasets are
    exchangeable.
}
\keyword{package}
