% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestspecies.R
\name{nearestspecies}
\alias{nearestspecies}
\title{Find nearest species compositional neighbors among two
    datasets}
\usage{
nearestspecies(spe1, spe2, method, ties = FALSE, ...)
}
\arguments{
\item{spe1}{species abundance matrix with greater number of sites
(rows)}

\item{spe2}{species abundance matrix with lesser number of sites
(rows)}

\item{method}{dissimilarity index, per
\code{\link[vegan]{vegdist}}; default uses Bray-Curtis
dissimilarities with stepacross adjustment}

\item{ties}{logical, should ties (non-unique values) be allowed?
default is \code{FALSE}}

\item{...}{additional arguments passed to function}
}
\value{
Numeric vector indexing rows (sample units) in \code{spe1} which
    are nearest compositional neighbors of \code{spe2}. The order
    corresponds to row order in \code{spe2}.
}
\description{
Given two species abundance matrices (of possibly
    unequal size), find nearest species compositional neighbors.
}
\details{
Useful to subset the longer of two datasets for analyses that
    require compatible dimensions (e.g.,
    \code{\link[vegan]{mantel}}, \code{\link[vegan]{procrustes}}).

    If \code{ties = FALSE}, then all returned values are unique;
    if \code{ties = TRUE}, then unique values are not enforced, so
    any sample unit in \code{spe1} may be picked repeatedly.
}
\examples{
# two candidate datasets, full and partial
set.seed(181)
data(smoky)
full <- smoky$spe
part <- (full + abs(rnorm(prod(dim(full)), 0, 25)))[1:6,]

# subset the full matrix, based on compositional nearest neighbors
i <- nearestspecies(full, part, ties=FALSE)
i
full <- full[i,,]

# default is to break ties
nearestspecies(full, part, ties=FALSE)
nearestspecies(full, part, ties=TRUE)

}
