% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_nmds.R
\name{recip_nmds}
\alias{recip_nmds}
\alias{summary.recip_nmds}
\alias{plot.recip_nmds}
\alias{plot_marg_grp}
\alias{plot_axismatch}
\title{Reciprocal NMDS}
\usage{
recip_nmds(tw, method = "bray", zeroadj = TRUE, step = TRUE, k = 2, ...)

\method{summary}{recip_nmds}(object, ...)

\method{plot}{recip_nmds}(x, type = "points", leg = FALSE, noaxes = TRUE,
  ...)

plot_marg_grp(object, ...)

plot_axismatch(object, ...)
}
\arguments{
\item{tw}{list of species and environment matrices, of class
\sQuote{twin}}

\item{method}{dissimilarity index, per \code{\link[vegan]{vegdist}}}

\item{zeroadj}{default \code{TRUE}: adjust dissimilarities to
account for zero-sum rows?}

\item{step}{default \code{TRUE}: use shortest-path stepacross
dissimilarity adjustment?}

\item{k}{number of dimensions sought in final NMDS solution}

\item{...}{additional arguments passed to function}

\item{x, object}{result from \code{recip_nmds}}

\item{type}{for plotting, one of \code{'points'}, \code{'text'},
or \code{'twin'}}

\item{leg}{logical, should legend be plotted?}

\item{noaxes}{default \code{TRUE}: cleanly plot without axes?}
}
\value{
List of class \sQuote{recip_nmds} with elements:\cr

\describe{
  \item{\code{m1}:}{
       Reciprocal model 1, calibrated with dataset 1 then swapping
       in dataset 2.)
  }
  \item{\code{m2}:}{
       Reciprocal model 2, calibrated with dataset 2 then swapping
       in dataset 1.)
  }
  \item{\code{grp}:}{
       Vector identifying group membership in dataset 1 or 2.
  }
  \item{\code{sumtab}:}{
       \code{data.frame} summary table, with elements:
\describe{
  \item{d1_rP, d2_rP}{
       \emph{Partial} intermodel fit for group 1, group 2 data
       (i.e., how well each dataset fits the opposing calibration
       model).
  }
  \item{rP_external}{
       \emph{Complete} intermodel fit (degree of overall external
       exchangebility among the two datasets, measured as
       Procrustean agreement of the two reciprocal models).
  }
  \item{stress1, stress2}{
       Respective stress of models 1,2.
  }
  \item{varexp1, varexp2}{
       Respective variance explained of models 1,2.
  }
  }
  }
  }
}
\description{
Reciprocal nonmetric multidimensional scaling (NMDS)
    to estimate external exchangeability of two datasets.
}
\details{
Reciprocal NMDS with \code{recip_nmds} estimates external
exchangeability of two candidate datasets.  The premise of
reciprocal NMDS is to calibrate each of two models with each
respective dataset, mutually exchanging datasets among the
alternative ordination models, then determining how well each
dataset fit the alternative model.  This is really just a special
case of 2-fold cross-validation as applied to NMDS.  The primary
use of reciprocal NMDS is to test the null hypothesis that two
multivariate datasets are exchangeable.
}
\examples{
# prepare two candidate datasets (here we just modify one)
set.seed(231)
data(smoky)
spe1 <- smoky$spe
env1 <- env2 <- smoky$env
spe2 <- spe1 + abs(rnorm(prod(dim(spe1)), 0, 2))  # add noise
tw   <- twin(spe1, spe2, env1, env2)

# reciprocal NMDS
r <- recip_nmds(tw)
summary(r)
plot(r)
plot(r, 'text')
plot(r, 'twin')
plot_marg_grp(r)
plot_axismatch(r)
plot_axismatch(r, ann=FALSE)

}
\references{
Kruskal, J. B. 1964. Multidimensional scaling by optimizing
  goodness of fit to a nonmetric hypothesis. Psychometrika 29:
  1-27.
}
\seealso{
\code{\link{resamp_nmds}} for resampled NMDS, and
     \code{\link{nearestspecies}} to force same number of rows
     among candidate datasets.
}
