% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_nmds.R
\name{resamp_nmds}
\alias{resamp_nmds}
\alias{summary.resamp_nmds}
\alias{plot.resamp_nmds}
\title{Resampled NMDS}
\usage{
resamp_nmds(spe, method = "bray", zeroadj = TRUE, step = TRUE,
  k = 2, BS = NULL, B = 9, ...)

\method{summary}{resamp_nmds}(object, ...)

\method{plot}{resamp_nmds}(x, noaxes = F, col = "#00000040",
  type = "spider", ...)
}
\arguments{
\item{spe}{species abundance matrix}

\item{method}{dissimilarity index, per \code{\link[vegan]{vegdist}}}

\item{zeroadj}{default \code{TRUE}: adjust dissimilarities to
account for zero-sum rows?}

\item{step}{default \code{TRUE}: use shortest-path stepacross
dissimilarity adjustment?}

\item{k}{number of dimensions sought in final NMDS solution}

\item{BS}{bootstrap size (currently constrained to not exceed the
number of rows in original matrix). Default \code{BS=NULL}, or
else \code{BS=1}, will perform jackknife NMDS (leave-one-out).}

\item{B}{number of bootstrap replicates}

\item{...}{additional arguments passed to function}

\item{x, object}{result from \code{resamp_nmds}}

\item{noaxes}{default \code{TRUE}: cleanly plot without axes?}

\item{col}{character vector of colors for spider lines}

\item{type}{plot type, one of \code{'spider'} or \code{'points'}
for ordination plots, or \code{'hist'} to plot the distribution
of fit statistics across resampling replicates}
}
\value{
List of class \sQuote{resamp_nmds} with elements:\cr
\describe{
     \item{\code{rP_draw}:}{
          Vector of rP_internal values for each replicate.
     }
     \item{\code{stresses}:}{
          Vector of NMDS stress for each replicate.
     }
     \item{\code{SRV}:}{
          Scaled rank variance across all replicates.
     }
     \item{\code{rP_internal}:}{
          Internal sampling variability across all replicates.
     }
     \item{\code{points}:}{
          Scores of each replicate configuration.
     }
}
}
\description{
Bootstrap or jackknife resampling of nonmetric
    multidimensional scaling (NMDS) configurations, to estimate
    internal sampling variability. Accompanied by summary and plot
    methods.
}
\details{
Bootstrap or jackknife resampling of NMDS with \code{resamp_nmds}
estimates internal sampling variability of a candidate dataset.
The premise of this is resampling with replacement from the
dissimilarities matrix, performing NMDS ordination (Kruskal 1964)
on each resampled replicate, then determining collective sampling
variability across all NMDS solutions as \code{rP_internal}.  Other
uses of resampled NMDS may include estimating confidence regions
for site scores, or testing the \sQuote{stability} of species
positions along ordination axes using Scaled Rank Variance of Knox
and Peet (1989).

Users may select dissimilarity index per
\code{\link[vegan]{vegdist}}, as well as options for zero-adjusted
dissimilarities (Clarke et al. 2006) using \code{zeroadj=TRUE},
and/or shortest-path \code{\link[vegan]{stepacross}}
dissimilarities using \code{step=TRUE}.
}
\examples{
set.seed(231)
data(smoky)
x <- smoky$spe

# try jackknife and bootstrap on a simplistic dataset
#    (vegan::metaMDS gives warning of near-zero stress)
j <- resamp_nmds(x, k=2)
b <- resamp_nmds(x, B=29, BS=NROW(x), k=2)
summary(j)
summary(b)
par(mfrow=c(1,2))
plot(j)
plot(b)

}
\references{
Clarke, K. R., P. J. Somerfield, and M. G. Chapman. 2006. On
    resemblance measures for ecological studies, including
    taxonomic dissimilarities and a zero-adjusted Bray–Curtis
    coefficient for denuded assemblages. J Exp Marine Biol and Ecol
    330:55–80.\cr
Knox, R. G., and R. K. Peet. 1989. Bootstrapped ordination: a
    method for estimating sampling effects in indirect gradient
    analysis. Vegetatio 80:153–165.\cr
Kruskal, J. B. 1964. Multidimensional scaling by optimizing
    goodness of fit to a nonmetric hypothesis. Psychometrika 29:
    1-27.\cr
}
\seealso{
\code{\link{recip_nmds}} for reciprocal NMDS, and
    \code{\link[vegan]{metaMDS}} for the core NMDS algorithm
}
\concept{resampled NMDS functions}
