% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin.R
\name{twin}
\alias{twin}
\alias{untwin}
\title{Twin/Untwin}
\usage{
twin(spe1, spe2, id1, id2, ...)

untwin(x, ...)
}
\arguments{
\item{spe1}{dataset 1 species matrix}

\item{spe2}{dataset 2 species matrix}

\item{id1}{dataset 1 environmental/identifier matrix}

\item{id2}{dataset 2 environmental/identifier matrix}

\item{...}{additional arguments passed to function}

\item{x}{twin object}
}
\value{
\code{twin} returns a list of two named items: `z1` and `z2`, each
    of which is a list with two named items: `spe` and `id` \cr
\code{untwin} returns a list with two named items: `spe` and `id`
}
\description{
Put twin datasets into one object (a list of lists),
    or else combine a twin object into a merged list.
}
\examples{
data(smoky)
spe <- smoky$spe
env <- smoky$env
tw  <- twin(spe, spe, env, env)
utw <- untwin(tw)

# must have same number of rows per dataset:
\dontrun{twin(spe, spe[3:9,], env, env[3:9,])}

}
\seealso{
\code{\link{nearestspecies}} to force same number of rows
}
