% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nmds.R
\name{boot_nmds}
\alias{boot_nmds}
\alias{summary.boot_nmds}
\alias{plot_boot_spider}
\title{Bootstrap NMDS}
\usage{
boot_nmds(x, B = 9, BS, k = 2, rot = TRUE, ...)

\method{summary}{boot_nmds}(object, ...)

plot_boot_spider(object, noaxes = F, col = rgb(0, 0, 0, 10, max = 255), ...)
}
\arguments{
\item{x}{list of species and environment matrices}

\item{B}{number of bootstrap replicates}

\item{BS}{bootstrap size (if not equal to number of rows in
original matrix)}

\item{k}{number of dimensions sought in final NMDS solution}

\item{rot}{logical, should solution be rotated to similarity with
an environmental gradient?}

\item{...}{additional arguments passed to function}

\item{object}{result from \code{boot_nmds}}

\item{noaxes}{logical, plot without axes?}

\item{col}{character vector of colors for spider lines}
}
\value{
List of class \sQuote{boot_nmds} with elements:
}
\description{
Bootstrap nonmetric multidimensional scaling (NMDS) to
    estimate internal stability.
}
\details{
Bootstrapped NMDS with \code{boot_nmds} estimates internal fit
(stability) of a candidate dataset.  The premise of bootstrapped
ordination is resampling with replacement from the reference
dataset, performing NMDS ordination on each bootstrap replicate,
then determining collective stability across all NMDS solutions.
Other uses of bootstrap NMDS may include estimating confidence
regions for ordination site scores, or testing the stability of
species positions along ordination axes.
}
\examples{
set.seed(231)
data(smoky)
x   <- list(spe=smoky$spe, id=smoky$env[,1:4])
res <- boot_nmds(x, B=29, k=2, rot=TRUE)
summary(res)
plot_boot_spider(res, col='#00000040')

# alternative plotting:
rn  <- gsub('\\\\.[[:alnum:]]+$', '', row.names(res$bootpt))
plot_boot_spider(res, col=ecole::colvec(as.factor(rn), alpha=0.5))

}
\references{
Kruskal, J. B. 1964. Multidimensional scaling by optimizing
  goodness of fit to a nonmetric hypothesis. Psychometrika 29:
  1-27.
}
\seealso{
\code{\link{recip_nmds}} for reciprocal NMDS, and
    \code{\link[vegan]{metaMDS}} for the core NMDS algorithm
}
