% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestspecies.R
\name{nearestspecies}
\alias{nearestspecies}
\title{Find nearest species compositional neighbors among two
    datasets}
\usage{
nearestspecies(spe1, spe2, method, uniq = FALSE, ...)
}
\arguments{
\item{spe1}{species abundance matrix with greater number of sites
(rows)}

\item{spe2}{species abundance matrix with lesser number of sites
(rows)}

\item{method}{dissimilarity index, per
\code{\link[vegan]{vegdist}}; default uses Bray-Curtis
dissimilarities with stepacross adjustment}

\item{uniq}{logical, force unique values? (currently unimplemented)}

\item{...}{additional arguments passed to function}
}
\value{
Numeric vector indexing sites in \code{spe1} which are nearest
    compositional neighbors of \code{spe2}.
}
\description{
Given two species abundance matrices (of possibly
    unequal size), find nearest species compositional neighbors.
}
\details{
Useful to subset the longer of two datasets for analyses that
    require compatible dimensions (e.g.,
    \code{\link[vegan]{mantel}}, \code{\link[vegan]{procrustes}}).
    Currently, any sample unit in \code{spe1} may be picked
    repeatedly (unique values not enforced).
}
\examples{
# two candidate datasets, full and partial
data(smoky)
full <- smoky$spe
part <- full[3:9,]

# subset the full matrix, based on compositional nearest neighbors
keep <- nearestspecies(full, part)
keep
full <- full[keep,,]

}
