% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_nmds.R
\name{recip_nmds}
\alias{recip_nmds}
\alias{summary.recip_nmds}
\alias{plot.recip_nmds}
\alias{plot_marg_grp}
\alias{plot_axismatch}
\title{Reciprocal NMDS}
\usage{
recip_nmds(x, k = 2, ...)

\method{summary}{recip_nmds}(object, ...)

\method{plot}{recip_nmds}(x, type = "points", leg = FALSE, noaxes = TRUE,
  ...)

plot_marg_grp(object, ...)

plot_axismatch(object, ...)
}
\arguments{
\item{x}{list of species and environment matrices, of class
\sQuote{twin}}

\item{k}{number of dimensions sought in final NMDS solution}

\item{...}{additional arguments passed to function}

\item{object}{result from \code{recip_nmds}}

\item{type}{for plotting, one of \code{c('points', 'text',
'twinned')}}

\item{leg}{logical, should legend be plotted?}

\item{noaxes}{logical, should a clean configuration with no axes be
plotted?}
}
\value{
List of class 'recip_nmds' with elements:
}
\description{
Reciprocal nonmetric multidimensional scaling (NMDS)
    to estimate external exchangeability of two datasets.
}
\details{
Reciprocal NMDS with \code{recip_nmds} estimates external fit
(exchangeability) of two candidate datasets.  The premise of
reciprocal NMDS is to calibrate each of two models with each
respective dataset, mutually exchanging datasets among the
alternative ordination models, then determining measures of how
well each dataset fit the alternative model.  This is really just a
special case of 2-fold cross-validation as applied to NMDS.  The
primary use of reciprocal NMDS is to test the null hypothesis that
two multivariate datasets are exchangeable.
}
\examples{
# prepare two candidate datasets (here we just modify one)
set.seed(231)
data(smoky)
spe1 <- smoky$spe
env1 <- env2 <- smoky$env
spe2 <- spe1 + abs(rnorm(prod(dim(spe1)), 0, 2))  # add noise
tw   <- twin(spe1, spe2, env1, env2)

# reciprocal NMDS
r <- recip_nmds(tw)
summary(r)
plot(r)
plot(r, 'text')
plot(r, 'twin')
plot_marg_grp(r)
plot_axismatch(r)
plot_axismatch(r, ann=FALSE)

}
\references{
Kruskal, J. B. 1964. Multidimensional scaling by optimizing
  goodness of fit to a nonmetric hypothesis. Psychometrika 29:
  1-27.
}
\seealso{
\code{\link{boot_nmds}} for bootstrap NMDS, and
     \code{\link{nearestspecies}} to force same number of rows
     among candidate datasets
}
